mem0fun/requirements.txt
mem0ai
qdrant-client
ollama
